requirements.txt
customtkinter
google-genai
pyperclip
python-dotenv
fpdf